refactor(api): split into different resources
* fix(api): account api admin auth
