+++ crates/interledger-api/src/routes/mod.rs
@@ -1,0 +1,9 @@
+mod accounts;
+mod ilp;
+mod settings;
+mod spsp;
+
+pub use accounts::AccountsApi;
+pub use ilp::IlpApi;
+pub use settings::SettingsApi;
+pub use spsp::SpspApi;
