+++ crates/interledger/Cargo.toml
@@ -60,7 +60,6 @@ parking_lot = "0.7.1"
 ring = "0.14.6"
 serde = "1.0.89"
 tokio = "0.1.20"
-tower-web = "0.3.7"
 url = "1.7.2"
 
 [dev-dependencies]
